refact: reorganize testing modules, move non-public code to `archive_testtools`.
